refactor. Create hierarchy of files for memory map.
