+++ cli/src/commands/memory-map-command.ts
@@ -1,7 +1,7 @@
 import type {EnvBuilder} from "../lib/env-builder";
 import {UpgradeChanges, UpgradeImporter, ZkSyncEraState} from "../lib/index";
 import {RpcClient} from "../lib/rpc-client";
-import {MemoryMap} from "../lib/memory-map";
+import {MemoryMap} from "../lib/memory-map/memory-map";
 import CliTable from "cli-table3";
 import chalk from "chalk";
