+++ cli/src/lib/memory-map/property-change.ts
@@ -1,0 +1,18 @@
+import {Property} from "./property";
+import {Option} from "nochoices";
+
+export class PropertyChange {
+  prop: Property
+  before: Option<string>
+  after: Option<string>
+
+  constructor (
+    prop: Property,
+    before: Option<string>,
+    after: Option<string>
+  ) {
+    this.prop = prop
+    this.before = before
+    this.after = after
+  }
+}
