+++ cli/src/lib/memory-map/types/blob-type.ts
@@ -1,0 +1,21 @@
+import type {MemoryDataType} from "./data-type";
+import {Option} from "nochoices";
+import {bytesToHex} from "viem";
+
+import {MemorySnapshot} from "../memory-snapshot";
+
+export class BlobType implements MemoryDataType {
+  extract (memory: MemorySnapshot, slot: bigint): Option<string> {
+    return memory.at(slot)
+      .map(this.format)
+      .map(str => str.toLowerCase())
+  }
+
+  format (data: Buffer): string {
+    return bytesToHex(data, {size: 32})
+  }
+
+  get evmSize (): number {
+    return 32;
+  }
+}
