+++ cli/src/lib/memory-map/types/data-type.ts
@@ -1,0 +1,9 @@
+import {Option} from "nochoices";
+
+import {MemorySnapshot} from "../memory-snapshot";
+
+export interface MemoryDataType {
+  extract (memory: MemorySnapshot, slot: bigint): Option<string>
+
+  get evmSize (): number
+}
